+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.17.0 (??)
 - Feature: new option to skip a page for later review
+- Feature: new option to display all the potential replacements
 - Detect protected pages
 - Detect redirection pages before analyzing the content
 - Cache pages to improve performance in particular on custom review